requirements.txt
Flask==2.0.1
flasgger==0.9.5
PyJWT==1.7.1  # Intentionally outdated
